Small refactoring; Further testbench improvements;
