+++ src/ScrollbarTrack.tsx
@@ -1,13 +1,8 @@
 import * as React from "react";
 import * as PropTypes from "prop-types";
 import cnb from "cnbuilder";
-import {
-  AXIS_DIRECTION,
-  AXIS_DIRECTION_PROP_TYPE,
-  ElementPropsWithElementRefAndRenderer,
-  renderDivWithRenderer
-} from "./common";
-import { isFun, isUndef } from "./util";
+import { AXIS_DIRECTION, AXIS_DIRECTION_PROP_TYPE, ElementPropsWithElementRefAndRenderer } from "./types";
+import { isFun, isUndef, renderDivWithRenderer } from "./util";
 
 export interface ScrollbarTrackClickParameters {
   axis: AXIS_DIRECTION;
