+++ testbench/app/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as ReactDom from "react-dom";
-import Scrollbar from "../../src/Scrollbar";
+import Scrollbar from "../../src";
 
 export const PARAGRAPHS_TEXT = [
   "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus aliquet malesuada efficitur. Praesent semper tortor id egestas volutpat. Aenean dui sapien, fermentum et dictum sagittis, finibus eget velit. Maecenas sed finibus risus, sed hendrerit odio. Nullam volutpat metus non enim consequat auctor. Vivamus gravida nibh in tempus vehicula. Donec venenatis luctus nulla, id facilisis turpis pharetra aliquet. Praesent non orci in turpis dapibus rutrum. Donec venenatis fermentum velit sit amet egestas. Duis quis ipsum et arcu scelerisque sagittis. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Aenean suscipit feugiat justo a luctus. Sed placerat sapien sit amet risus efficitur, sed pharetra tellus faucibus.",
@@ -21,7 +21,7 @@ export function getRandomParagraphText() {
   return PARAGRAPHS_TEXT[Math.floor(Math.random() * PARAGRAPHS_TEXT.length)];
 }
 
-export function renderAmountOfParagraphs(amount = 5, paragraphsProps = undefined) {
+export function renderAmountOfParagraphs(amount = 5, paragraphsProps: React.HTMLProps<HTMLParagraphElement> = {}) {
   const result: Array<JSX.Element> = [];
 
   for (; amount--; ) {
@@ -36,6 +36,15 @@ export function renderAmountOfParagraphs(amount = 5, paragraphsProps = undefined
 }
 
 ReactDom.render(
-  <Scrollbar style={{ width: 528, height: 200 }}>{renderAmountOfParagraphs(10)}</Scrollbar>,
+  <Scrollbar
+    style={{ width: 528, height: 200 }}
+    scrollerProps={{
+      elementRef: ref => {
+        console.log(ref);
+      }
+    }}
+  >
+    {renderAmountOfParagraphs(10, { style: { width: "150%" } })}
+  </Scrollbar>,
   document.getElementById("AppRoot")
 );
